+++ app/src/main/java/com/example/project_enlishlearning/viewmodel/VocabularyViewModel.kt
@@ -140,7 +140,7 @@ class VocabularyViewModel(application: Application) : AndroidViewModel(applicati
     }
 
     fun updateWord(
-        wordId: Int,
+        wordId: Long,
         setId: Int,
         word: String,
         pronunciation: String,
@@ -164,7 +164,7 @@ class VocabularyViewModel(application: Application) : AndroidViewModel(applicati
         }
     }
 
-    fun loadWordById(wordId: Int) {
+    fun loadWordById(wordId: Long) {
         viewModelScope.launch {
             _currentEditWord.value = dao.getWordById(wordId)
         }
